Retrieve claims from JWT token
