+++ src/libraries/Hosting/AspNetCore/CloudAdapter.cs
@@ -13,8 +13,6 @@ using Microsoft.Agents.Builder;
 using System.Text;
 using Microsoft.Agents.Core.Errors;
 using Microsoft.Agents.Hosting.AspNetCore.BackgroundQueue;
-using System.Linq;
-using System.IdentityModel.Tokens.Jwt;
 using Microsoft.Extensions.Configuration;
 
 namespace Microsoft.Agents.Hosting.AspNetCore
@@ -129,24 +127,7 @@ namespace Microsoft.Agents.Hosting.AspNetCore
                 // Deserialize the incoming Activity
                 var activity = await HttpHelper.ReadRequestAsync<IActivity>(httpRequest).ConfigureAwait(false);
 
-                // If Auth is not configured, we still need the claims from the JWT token.
-                // Currently, the stack does rely on certain Claims.  If the Bearer token
-                // was sent, we can get them from there.  The JWT token is NOT validated though.
-                var claimsIdentity = (ClaimsIdentity)httpRequest.HttpContext.User.Identity;
-                if (!claimsIdentity.IsAuthenticated && !claimsIdentity.Claims.Any())
-                {
-                    var auth = httpRequest.Headers.Authorization;
-                    if (auth.Count != 0)
-                    {
-                        var authHeaderValue = auth.First();
-                        var authValues = authHeaderValue.Split(' ');
-                        if (authValues.Length == 2 && authValues[0].Equals("bearer", StringComparison.OrdinalIgnoreCase))
-                        {
-                            var jwt = new JwtSecurityToken(authValues[1]);
-                            claimsIdentity = new ClaimsIdentity(jwt.Claims);
-                        }
-                    }
-                }
+                var claimsIdentity = HttpHelper.GetClaimsIdentity(httpRequest);
 
                 if (!IsValidChannelActivity(activity))
                 {
